+++ client/drizzledump.c
@@ -75,7 +75,7 @@
 
 static void add_load_option(DYNAMIC_STRING *str, const char *option,
                              const char *option_value);
-static ulong find_set(TYPELIB *lib, const char *x, uint length,
+static uint32_t find_set(TYPELIB *lib, const char *x, uint length,
                       char **err_pos, uint *err_len);
 
 static void field_escape(DYNAMIC_STRING* in, const char *from);
@@ -98,7 +98,7 @@ static bool  verbose= 0, opt_no_create_info= 0, opt_no_data= 0,
                 opt_events= 0,
                 opt_alltspcs=0, opt_notspcs= 0;
 static bool insert_pat_inited= 0, debug_info_flag= 0, debug_check_flag= 0;
-static ulong opt_max_allowed_packet, opt_net_buffer_length;
+static uint32_t opt_max_allowed_packet, opt_net_buffer_length;
 static DRIZZLE drizzle_connection,*drizzle=0;
 static DYNAMIC_STRING insert_pat;
 static char  *opt_password=0,*current_user=0,
@@ -112,7 +112,7 @@ static char **defaults_argv= 0;
 static char compatible_mode_normal_str[255];
 /* Server supports character_set_results session variable? */
 static bool server_supports_switching_charsets= true;
-static ulong opt_compatible_mode= 0;
+static uint32_t opt_compatible_mode= 0;
 #define DRIZZLE_OPT_MASTER_DATA_EFFECTIVE_SQL 1
 #define DRIZZLE_OPT_MASTER_DATA_COMMENTED_SQL 2
 #define DRIZZLE_OPT_SLAVE_DATA_EFFECTIVE_SQL 1
@@ -138,7 +138,7 @@ static void dynstr_append_checked(DYNAMIC_STRING* dest, const char* src);
 static void dynstr_set_checked(DYNAMIC_STRING *str, const char *init_str);
 static void dynstr_append_mem_checked(DYNAMIC_STRING *str, const char *append,
 			  uint length);
-static void dynstr_realloc_checked(DYNAMIC_STRING *str, ulong additional_size);
+static void dynstr_realloc_checked(DYNAMIC_STRING *str, uint32_t additional_size);
 /*
   Constant for detection of default value of default_charset.
   If default_charset is equal to drizzle_universal_client_charset, then
@@ -753,7 +753,7 @@ get_one_option(int optid, const struct my_option *opt __attribute__((unused)),
       char buff[255];
       char *end= compatible_mode_normal_str;
       int i;
-      ulong mode;
+      uint32_t mode;
       uint err_len;
 
       opt_quoted= 1;
@@ -1255,7 +1255,7 @@ static char *quote_for_like(const char *name, char *buff)
     Quote '<' '>' '&' '\"' chars and print a string to the xml_file.
 */
 
-static void print_quoted_xml(FILE *xml_file, const char *str, ulong len)
+static void print_quoted_xml(FILE *xml_file, const char *str, uint32_t len)
 {
   const char *end;
 
@@ -1438,7 +1438,7 @@ static void print_xml_row(FILE *xml_file, const char *row_name,
     Print hex value for blob data.
 */
 
-static void print_blob_as_hex(FILE *output_file, const char *str, ulong len)
+static void print_blob_as_hex(FILE *output_file, const char *str, uint32_t len)
 {
     /* sakaik got the idea to to provide blob's in hex notation. */
     const char *ptr= str, *end= ptr + len;
@@ -2050,7 +2050,7 @@ static void dump_table(char *table, char *db)
   char table_type[NAME_LEN];
   char *result_table, table_buff2[NAME_LEN*2+3], *opt_quoted_table;
   int error= 0;
-  ulong         rownr, row_break, total_length, init_length;
+  uint32_t         rownr, row_break, total_length, init_length;
   uint num_fields;
   DRIZZLE_RES     *res;
   DRIZZLE_FIELD   *field;
@@ -2279,7 +2279,7 @@ static void dump_table(char *table, char *db)
       for (i= 0; i < drizzle_num_fields(res); i++)
       {
         int is_blob;
-        ulong length= lengths[i];
+        uint32_t length= lengths[i];
 
         if (!(field= drizzle_fetch_field(res)))
           die(EX_CONSCHECK,
@@ -2432,7 +2432,7 @@ static void dump_table(char *table, char *db)
 
       if (extended_insert)
       {
-        ulong row_length;
+        uint32_t row_length;
         dynstr_append_checked(&extended_row,")");
         row_length= 2 + extended_row.length;
         if (total_length + row_length < opt_net_buffer_length)
@@ -2470,7 +2470,7 @@ static void dump_table(char *table, char *db)
     if (drizzle_errno(drizzle))
     {
       snprintf(buf, sizeof(buf),
-               "%s: Error %d: %s when dumping table %s at row: %ld\n",
+               "%s: Error %d: %s when dumping table %s at row: %d\n",
                my_progname,
                drizzle_errno(drizzle),
                drizzle_error(drizzle),
@@ -3128,11 +3128,11 @@ static int start_transaction(DRIZZLE *drizzle_con)
 }
 
 
-static ulong find_set(TYPELIB *lib, const char *x, uint length,
+static uint32_t find_set(TYPELIB *lib, const char *x, uint length,
                       char **err_pos, uint *err_len)
 {
   const char *end= x + length;
-  ulong found= 0;
+  uint32_t found= 0;
   uint find;
   char buff[255];
 
@@ -3404,7 +3404,7 @@ static void dynstr_append_mem_checked(DYNAMIC_STRING *str, const char *append,
     die(EX_DRIZZLEERR, DYNAMIC_STR_ERROR_MSG);
 }
 
-static void dynstr_realloc_checked(DYNAMIC_STRING *str, ulong additional_size)
+static void dynstr_realloc_checked(DYNAMIC_STRING *str, uint32_t additional_size)
 {
   if (dynstr_realloc(str, additional_size))
     die(EX_DRIZZLEERR, DYNAMIC_STR_ERROR_MSG);
